handshake: Retry on uTP timeout even when encryption disabled
Also remove redundancy in the logic a bit. The logic refactorings
should be a noop, so the only actual change should be that
we now additionally fallback from uTP to TCP if we timeout at the
AWAITING_HANDSHAKE stage on an outgoing connection.

Note that this can actually occur two ways: if we started by sending out a plaintext
handshake (which happens if we have TR_CLEAR_PREFERRED set), or if we completed
an encrypted handshake and sent out our message but received no reply.
The latter is a case in which there's technically no point retrying,
although it doesn't really hurt.