handshake: Fix connection between TR_ENCRYPTION_PREFERRED and peer which only supports header-encrypt.
When using TR_ENCRYPTION_PREFERRED mode we should offer both
message-encrypt and header-encrypt options. Our peer may only support
header-encrypt (e.g. maybe it's a low powered device but still uses
encrypted handshake to avoid traffic shaping). The other party will
choose the best out of these.

This is more in the spirit of "encryption _preferred_", as having
header-only encryption is still better than falling back to plaintext
handshake.

If you wanted to be really fancy, you'd actually define the fallbacks differently
for clear-preferred vs encryption-preferred. E.g. clear-preferred would try an outgoing
plaintext handshake first, and if that fails then attempt an encrypted handshake.
But that's too much complexity for the marginal benefit it gives (really
the only reason someone would be a stickler for plaintext is avoiding any CPU overhead
of encryption. But this is only really an issue for boatloads of torrents,
and most likely in such a case you are seeding as opposed to downloading, in which case
we are concerned with incoming connections and we get to determine the crypto protocol
anyhow.)