fix: fail to parse hybrid torrents due to zero-length dictionary key
Backport 72a671e7d07cc7f58fcfa54e786ee7cf1af0f0ff

The key issue appears to be that previously the code used tr_quark = 0
to signal when no key is currently present. But tr_quark = 0 actually
has a valid meaning as an empty string, which can legitimately appear
for hybrid torrents, and needs to be written out/should not be skipped.

So we need to choose some other way to explicitly indicate presence.
The official PR (4.0+ / C++) used std::optional for this purpose.
There was a 3.00-compatible (C variant) patch posted in the corresponding
issue, which added an additional bool flag for the presence.

The approach in this commit uses tr_quark = -1 (which is actually
uint64_t::max since tr_quark is of type size_t) to handle this.

This is a non-standard usage of tr_quark which conventionally only
takes on values in a specified enum range, however since its use
is localized and does not escape from the two parsing functions I think
it is fine.

Preferably should test with the .torrent file linked in the issue.