handshake: return READ_NOW after handshake success
Backport 1bedccea5853d0383c7592bfab34e01ce4e900e8

Without this, BT messages immediately following a
successful handshake that is then disconnected by peer
would never be processed. This can happen e.g. when peer is also a
seeder.

This happened because of the READ_LATER return value. When peer-io
sees this it retries later trying to pull data from a socket
that's now disconnected, and it errors out without processing
any data that might have been pending.

Fix this by changing return value to READ_NOW. Importantly
this value must be serviced after the canRead handler is swapped out,
so it must bubble out to peer-io (see the following paragraph).

Note that the loop in canread is a bit subtle.

The additional bit was originally added back in
4a8e34a35d32f4186ec4cea4cf789cbd59fa1c4b as some sort of hail-mary
trying to fix another bug. It wasn't even the right fix, but it is
actually useful.

In mainline transmission,

75dca6ecba7d7d053e0305713195c71e08736ef0
c01110d49ae6cba268781f4351d019a289df7a98

got rid of the addendum, which had the effect of looping whenever read_now
was true instead of only when
read_now == true AND we are in one ot the awaiting pad states with
a ready in buffer.

That was mostly OK, since each of the individual cases take care to
check whether or not stuff is in the buffer. (If it isn't, it returns
READ_LATER which is the proper behavior.) However in the case where
we completed a handshake, it's not appropriate to loop if READ_NOW is
returned. This is what the second commit above tries to fix, since upon
completing a handshake we want to start reading actual data.

This canRead loop inside handshake is itself wrapped inside an outer
loop within peer-op, so it is safe to remove any loop from inside
it.

So upstream transmission ultimately did get to the correct state.
But as noted, removing the loop entirely means they round-trip through
peer-io's function every time which has some logic to account for
the number of bytes (notify_bandwidth_consumed). This is avoided
by this approach.