peer-io: Clear read/write buffer when reconnecting
Backport 4a778bcb57c9a38e7b7d17f8b37d76faaa00c5a8

Otherwise there can be garbage stuck in the buffer when we
reconnect which can cause issues establishing an encrypted connection.
This was exhibited when for instance
DH keys are exchanged but torrent is not found (or is not running)
by the peer. At this point we have some "garbage" values stuck
in the buffer.