handshake: Fix peer connect fallback again
41a4742543254ed62cdf478a9602bfb72bf05e18 originally fixed things
so that a failed plain-text uTP handshake is correctly retried
as a plain-text TCP handshake.

But the subsequent refactoring in 6b9e08a2e16b499e87b55a2570d6613a4dcf203b
removed this logic. Maybe I'm just going in circles or something.

I think the context was that since as of 520d8af9378c4b3023adc98ae4bfe2a935390523
we always send out an encrypted handshake even in cleartext mode,
a plain-text uTP handshake is not a scenario that should happen. But
that's a non-obvious invariant, and it's clearer to just handle all
states properly here rather than making assumptions about the state
we can be in.

Hopefully this covers all the cases.