crypto: Clear out enc/dec keys on connection reset.
Looking at the handshake/crypto code a bit more, it seems that there
are cases where the mode is set to rc4 in peer-io but the crypto
state is still uninitialized (it starts off memset to 0, but
when we re-use the state for a reconnect it would have old values).

The previous code detected the case of enc/dec key NULL and
special-cased it. This is technically a logical bug, and there's even a fixme
in crypto.c for this. But for now the practical implication is that
it's not sufficient to just reset the encryption mode to plaintext,
we also have to reset the keys to clear out any residual encryption
state.

To do this in a clean way, we make the following changes

1) Introduce a new function (tr_cryptoReset) to clear out any state
   that should be reset, and call it on socket reconnect.

2) Fix a memory-leak when re-assigning crypto->mySecret in
   tr_cryptoComputeSecret.

3) Move the NULL-dec key hack from crypto to peer-io.