handshake: Refactor to centralize setting of peer crypto state.
Previously there was some messy global state tracking, complicated
by the fact that the handshake set peer-io as RC4 but kept the dec/enc
key uninitialized.

We refactor things to obey a uniform rule:
* Handshake starts off with messages as unencrypted
* We always set the enc/dec state before switching to encrypted

https://css.csail.mit.edu/6.5660/2018/projects/bgu-kelvinlu.pdf
was used as a reference for the MSE handshake process.

High chance of regressions or breakages with this.