handshake: Tweak offered crypto provide for plaintext
Now that I realize upstream Tr3 (and older) versions don't properly
support downgrading crypto, I realize that sending out an MSE
handshake that offers a plaintext option is a bad idea.

If our peer is a mainline Tr3 client in clear_preferred mode, it
will always prefer to negotiate crypto2 as cleartext. But it doesn't
handle downgrades properly, so that peer is going to be effectively
unreachable.

We should just do what mainline Tr3 does and ever only offer rc4
encryption as an option if we do an outgoing MSE handshake.

This does mean that like mainline, a client in clear_preferred
will never be able to make an outgoing connection to an encrypted_only
client.

So we basically pick between two poisons:
* Either our clear_preferred behavior is to make MSE offering plaintext,
   which will prevent it from proprerly connecting with mainline Tr3
   clear_preferred peer.

* Or our clear_preferred behavior is to make a plaintext handshake,
   which will prevent it from properly connecting with any encrypted_only
   peer.

I'm just going to go with the second option since it's at least somewhat
intuitive.