net: Fix wrong ipv6 used for announces
From vuori:
"tr_globalIPv6() will always return the system's idea of the "default"
outgoing address, let's call that address A. This address is the one that
the kernel will select based on a set of rules if the user has not
explicitly selected an outbound address by binding the socket before
calling connect or sendmsg.
Getting the result of the kernel's rule-processing is the idea behind
get_source_address.

If bind-address-ipv6 is set to something other than A, let's call this address B,
rebind_ipv6 will (correctly) listen on address B. But the HTTP and LTEP
announce functions will still advertise address A, where no one is
listening (because all inbound sockets are bound to B), and which the
user may not want to make public.

If user sets bind-address-ipv6 to something non-routable, that's
generally a user error and can't really be helped. The current behavior
may be correct in an obscure NAT situation, such as where the system
has routable address A and another address B which becomes NAT'd to A
somewhere down the line, but I have no idea why anyone would do this.
In other situations, the current behavior will generally advertise a
non-working address and also leak information."

Backport 20421876a57e44fa088bbe76e67a97da1c5b6cd6