Revert "TorrentTableView: Avoid use of private-method when fixing glitch"
This reverts commit e10490abd6bb6f8731dedd589bd151d229a00b17.

Apparently in some rare cases the setting of layer-backed still ends
up occurring too early before layout properly initialized. The end
result is the UI becomes corrupted.

Switch back to the previous approach of using a private method. This
seems to always be "safe" in that even if it's done too early it doesn't
corrupt anything.